+++ src/components/DrinkCard/DrinkCard.module.css
@@ -1,0 +1,33 @@
+@media screen and (max-width: 767px) {
+	.mainPageList_item:not(:first-child) {
+		display: none;
+		opacity: 0;
+		visibility: hidden;
+		pointer-events: none;
+	}
+}
+
+@media screen and (max-width: 1439px) {
+	.mainPageList_item:nth-child(n + 3) {
+		display: none;
+		opacity: 0;
+		visibility: hidden;
+		pointer-events: none;
+	}
+}
+
+.card_link {
+	display: inline-block;
+}
+
+.card_name {
+	font-size: 18px;
+}
+
+.card_text_wrapper {
+	display: flex;
+	flex-direction: row;
+	justify-content: space-between;
+
+	margin-top: 14px;
+}
